requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
